# vostory-server/docs/swagger.yaml
definitions:
...
  iot-alert-center_api_v1.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nunu Example API
  version: 1.0.0
paths:
  /api/v1/system/api/list:
    get:
      consumes:
      - application/json
      description: 分页获取系统API列表
      operationId: system:api:list
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      - description: 请求方法
        in: query
        name: method
        type: string
      - description: API路径
        in: query
        name: path
        type: string
      - description: API名称
        in: query
        name: name
        type: string
      - description: API描述
        in: query
        name: desc
        type: string
      - description: API标签
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
      summary: 获取API列表
      tags:
      - API管理
  /api/v1/system/api/tag/list:
    get:
      consumes:
      - application/json
      description: 获取系统API标签列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
      summary: 获取API标签列表
      tags:
      - API管理
...
  /api/v1/system/logininfor/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取登录日志详情
      operationId: system:logininfor:detail
      parameters:
      - description: 登录日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
      summary: 获取登录日志详情
      tags:
      - 登录日志
  /api/v1/system/logininfor/list:
    get:
      consumes:
      - application/json
      description: 分页获取登录日志列表
      operationId: system:logininfor:list
      parameters:
      - description: 当前页
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      - description: 登录账号
        in: query
        name: login_name
        type: string
      - description: 登录IP地址
        in: query
        name: ip_addr
        type: string
      - description: 登录状态
        in: query
        name: status
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/iot-alert-center_api_v1.Response'
      summary: 获取登录日志列表
      tags:
      - 登录日志
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"